requirements.txt
pandas
pydantic
plotly
streamlit
openpyxl
python-dotenv